requirements.txt
discord.py==1.7.2
python-dotenv==0.18.0